+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 _nothing changed yet_
 
 
+## [v0.1.1] - 2025-10-24
+
+### Changed
+
+- Property names of the event should be more descriptive:
+  - From `CallerNumber` to `ExternalNumber`
+  - From `CalleeNumber` to `InternalNumber`
+
+
 ## [v0.1.0] - 2025-08-28
 
 _Inital release_
@@ -27,6 +36,7 @@ _Inital release_
 
 
 
-[Unreleased]: https://github.com/AM-WD/FritzCallMonitor/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/AM-WD/FritzCallMonitor/compare/v0.1.1...HEAD
 
+[v0.1.1]: https://github.com/AM-WD/FritzCallMonitor/compare/v0.1.0...v0.1.1
 [v0.1.0]: https://github.com/AM-WD/FritzCallMonitor/commits/v0.1.0